+++ packages/lynjs/src/element/symbol.ts
@@ -2,4 +2,4 @@ export const CONTROLLER: unique symbol = Symbol('controller');
 export const CONNECTED: unique symbol = Symbol('connected');
 export const ATTRIBUTES: unique symbol = Symbol('attributes');
 export const STATE: unique symbol = Symbol('state');
-export const HOST: unique symbol = Symbol('host');
+export const HOST: unique symbol = Symbol('hostNode');